# api/openapi.yaml
openapi: 3.0.3
info:
  title: Runnix
  version: 0.1.0
  description: >
    Runnix API. All executions and tenants routes require
    Authorization: Bearer <jwt> AND X-Tenant-ID. 401 = bad token,
    400 = missing tenant, 403 = not a member.
servers:
  - url: http://localhost:4000
paths:
  /healthz:
    get:
      summary: Liveness
      responses: { '200': { description: ok } }
  /readyz:
    get:
      summary: Readiness (live Postgres + NATS checks)
      responses:
        '200': { description: ready }
        '503': { description: a dependency is unavailable }
  /languages:
    get:
      summary: Supported languages
      responses: { '200': { description: ok } }
  /auth/register:
    post:
      summary: "Register"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string, minLength: 3 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: created; personal tenant auto-provisioned, token pair issued
  /auth/login:
    post:
      summary: "Login"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                username: { type: string }
                password: { type: string }
      responses:
        '200': { description: ok, token pair issued }
  /auth/refresh:
    post:
      summary: "Refresh"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: ok, rotated token pair issued }
  /tenants:
    post:
      summary: Create org tenant (caller becomes owner)
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 3, maxLength: 60 }
                slug: { type: string, pattern: '^[a-z0-9-]{3,40}$', description: optional; derived from name when absent }
      responses:
        '201': { description: tenant created }
        '400': { description: invalid name/slug }
        '403': { description: authentication required (no user identity) }
  /tenants/{id}:
    get:
      summary: Get tenant (id must equal X-Tenant-ID)
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: ok }
        '403': { description: id does not match X-Tenant-ID or not a member }
        '404': { description: tenant not found }
  /executions:
    post:
      summary: Submit execution (persisted queued, published to JetStream)
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, language, source]
              properties:
                tenant_id: { type: string, format: uuid, description: must equal X-Tenant-ID }
                language: { type: string, enum: [python] }
                source: { type: string, maxLength: 102400 }
                stdin: { type: string, maxLength: 102400 }
                timeout_s: { type: integer, minimum: 1, maximum: 60, default: 2 }
                webhook_url: { type: string, format: uri, description: http(s) only for now; SSRF blocklist deferred }
      responses:
        '202': { description: 'queued; body {id, status, created_at}' }
        '400': { description: invalid input }
        '403': { description: tenant_id does not match X-Tenant-ID }
        '502': { description: submit queue unavailable; execution marked failed }
    get:
      summary: List executions for tenant (newest first)
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200': { description: 'ok; body {executions, limit, offset}' }
  /executions/{id}:
    get:
      summary: Get execution (always tenant-scoped; 404 when owned by another tenant)
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 'ok; full execution incl. source, stdout, stderr, exit_code' }
        '400': { description: invalid id }
        '404': { description: execution not found }
components:
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
security:
  - bearerAuth: []